refactor: use shared theme components and drop forked copies
Point FirstContactTheme, AftermathTheme and SideComponents at the
consolidated theme/shared components (passing variant="aftermath" where
needed) and delete the eight near-duplicate per-theme copies of NewsFeed,
BreakingNews, LogoBlock and CurrentTurn.
