+++ src/components/SideComponents.tsx
@@ -5,12 +5,12 @@ import {
     SetupInformation,
 } from "@fc/types/types";
 import Icon_Game from "@fc/public/Icon-VLHG.png";
-import LogoBlock from "./theme/first-contact/LogoBlock";
+import LogoBlock from "./theme/shared/LogoBlock";
 import React, { useRef, useState } from "react";
 import { SideComponentMapper } from "@fc/lib/ComponentMapper";
 import { CSSTransition, SwitchTransition } from "react-transition-group";
 import useInterval from "@fc/lib/useInterval";
-import { NewsFeed } from "@fc/components/theme/first-contact/NewsFeed";
+import { NewsFeed } from "@fc/components/theme/shared/NewsFeed";
 import TabSwitcher, { TabInfo } from "@fc/components/TabSwitcher";
 import { calculatePressTabIcon } from "@fc/lib/press";
