+++ src/components/theme/first-contact/FirstContactTheme.tsx
@@ -3,11 +3,11 @@ import { useState } from "react";
 import { GameTabSwitcher } from "../../TabSwitcher";
 import SideComponentWrapper from "../../SideComponents";
 import { SideComponentMapper } from "@fc/lib/ComponentMapper";
-import BreakingNews from "./BreakingNews";
-import { NewsFeed } from "./NewsFeed";
-import LogoBlock from "./LogoBlock";
+import BreakingNews from "../shared/BreakingNews";
+import { NewsFeed } from "../shared/NewsFeed";
+import LogoBlock from "../shared/LogoBlock";
 import TurnCounter from "./TurnCounter";
-import CurrentTurn from "./CurrentTurn";
+import CurrentTurn from "../shared/CurrentTurn";
 import { getCurrentPhase } from "@fc/server/turn";
 import { isLeft } from "fp-ts/Either";
 import PhaseInformation from "@fc/components/PhaseInformation";
